Fix for arm IK jitter
When in HMD mode, holding your hand controller in front of your face, would often result in a 1-2 mm pop in wrist position.
This was due to a problem in the IK elbow constraint.
When the swing rotation out of the elbow's reference frame was close to but not exactly identity it would not be properly constrainted.
But when the magnitude of that swing became big enough the constraint would engage.
This small change in swing rotation was noticable as a 1-2 mm movement in the wrist.

(cherry picked from commit f6c4f6b9c4aa9494458244f7a752de884b66a6f4)