Small performance improvement in MyAvatar::getJointIndex()
While investigating a crash, I noticed that this code is far from optimal.
It can perform unnecessary memory allocation and does an unnecessary hash lookup.

From QT's QHash documentation:

http://doc.qt.io/qt-5/qhash.html

In general, we recommend that you use `contains()` and `value()`
rather than `operator[]()` for looking up a key in a hash. The
reason is that `operator[]()` silently inserts an item into the
hash if no item exists with the same key (unless the hash is
const). For example, the following code snippet will create 1000
items in memory: